generate the pairs from metadata
importer imports the pairs that univocity knows about and that are supported by the exchange
