+++ univocity-trader-core/src/test/java/com/univocity/trader/currency/TickersTest.java
@@ -4,6 +4,8 @@ import static junit.framework.TestCase.assertTrue;
 
 import org.junit.Test;
 
+import com.univocity.trader.tickers.Tickers;
+
 public class TickersTest {
    @Test
    public void testTickersLoad() {
